# Conflicts:
#	README.md
#	app/access_controls/access_control_nodes.php
#	app/basic_operator_panel/resources/content.php
#	app/bridges/bridge_delete.php
#	app/bridges/bridge_edit.php
#	app/call_broadcast/call_broadcast_delete.php
#	app/call_centers/call_center_queue_edit.php
#	app/conference_controls/app_defaults.php
#	app/conferences_active/conferences_active_inc.php
#	app/contacts/contact_addresses.php
#	app/contacts/contact_edit.php
#	app/contacts/contact_time_edit.php
#	app/contacts/contacts_vcard.php
#	app/destinations/destinations.php
#	app/devices/device_edit.php
#	app/dialplans/dialplans.php
#	app/edit/clip_options.php
#	app/edit/filedelete.php
#	app/edit/filenew.php
#	app/edit/filerename.php
#	app/edit/filesave.php
#	app/edit/folderdelete.php
#	app/edit/foldernew.php
#	app/edit/index.php
#	app/email_templates/app_defaults.php
#	app/extensions/extension_copy.php
#	app/extensions/extension_download.php
#	app/fifo_list/fifo_interactive_inc.php
#	app/fifo_list/fifo_list_inc.php
#	app/ivr_menus/app_defaults.php
#	app/ivr_menus/ivr_menu_edit.php
#	app/pin_numbers/pin_download.php
#	app/polycom/app_config.php
#	app/provision/resources/classes/provision.php
#	app/ring_groups/app_defaults.php
#	app/ring_groups/ring_group_edit.php
#	core/authentication/resources/classes/plugins/database.php
#	core/authentication/resources/classes/plugins/ldap.php
#	core/domains/app_languages.php
#	core/domains/domain_edit.php
#	core/users/app_defaults.php
#	core/users/user_edit.php
#	core/users/users.php
#	resources/functions.php
#	resources/templates/provision/polycom/5.x/{$mac}.cfg
#	themes/default/template.php

+++ app/destinations/destinations.php
@@ -124,6 +124,7 @@
 	}
 	$page = $_GET['page'];
 	if (strlen($page) == 0) { $page = 0; $_GET['page'] = 0; }
+	list($paging_controls_mini, $rows_per_page, $var_3) = paging($num_rows, $param, $rows_per_page, true);
 	list($paging_controls, $rows_per_page, $var3) = paging($num_rows, $param, $rows_per_page);
 	$offset = $rows_per_page * $page;
 
@@ -211,6 +212,9 @@
 
 	echo "				<input type='text' class='txt' style='width: 150px; margin-left: 15px;' name='search' id='search' value='".escape($search)."'>\n";
 	echo "				<input type='submit' class='btn' name='submit' value='".$text['button-search']."'>\n";
+	if ($paging_controls_mini != '') {
+		echo "			<span style='margin-left: 15px;'>".$paging_controls_mini."</span>\n";
+	}
 	echo "			</td>\n";
 	echo "		</form>\n";
 	echo "	</tr>\n";
@@ -233,7 +237,9 @@
 	echo th_order_by('destination_type', $text['label-destination_type'], $order_by, $order, $param);
 	echo th_order_by('destination_number', $text['label-destination_number'], $order_by, $order, $param);
 	echo  "<th>". $text['label-detail_action']."</th>";
-	echo th_order_by('destination_context', $text['label-destination_context'], $order_by, $order, $param);
+	if (permission_exists("destination_context")) {
+		echo th_order_by('destination_context', $text['label-destination_context'], $order_by, $order, $param);
+	}
 	if (permission_exists('outbound_caller_id_select')) {
 		echo th_order_by('destination_caller_id_name', $text['label-destination_caller_id_name'], $order_by, $order, $param);
 		echo th_order_by('destination_caller_id_number', $text['label-destination_caller_id_number'], $order_by, $order, $param);
@@ -265,7 +271,7 @@
 			echo "	</td>\n";
 			if ($_GET['show'] == "all" && permission_exists('destination_all')) {
 				if (strlen($_SESSION['domains'][$row['domain_uuid']]['domain_name']) > 0) {
-					$domain = escape($_SESSION['domains'][$row['domain_uuid']]['domain_name']);
+					$domain = $_SESSION['domains'][$row['domain_uuid']]['domain_name'];
 				}
 				else {
 					$domain = $text['label-global'];
@@ -274,9 +280,11 @@
 			}
 			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_type'])."&nbsp;</td>\n";
 			echo "	<td valign='top' class='".$row_style[$c]."'>".escape(format_phone($row['destination_number']))."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($action_name)."&nbsp;</td>\n";
+			echo "	<td valign='top' class='".$row_style[$c]."'>".$action_name."&nbsp;</td>\n";
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".$row['destination_number_regex']."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_context'])."&nbsp;</td>\n";
+			if (permission_exists("destination_context")) {
+				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_context'])."&nbsp;</td>\n";
+			}
 			//echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['fax_uuid'])."&nbsp;</td>\n";
 			if (permission_exists('outbound_caller_id_select')) {
 				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['destination_caller_id_name'])."&nbsp;</td>\n";
@@ -294,7 +302,7 @@
 				echo "<a href='destination_edit.php?id=".escape($row['destination_uuid'])."' alt='".$text['button-edit']."'>$v_link_label_edit</a>";
 			}
 			if (permission_exists('destination_delete')) {
-				echo "<button type='submit' class='btn btn-default list_control_icon' name=\"destinations[$x][action]\" alt='".$text['button-delete']."' value='delete'><span class='glyphicon glyphicon-remove'></span></button>";
+				echo "<button type='submit' class='btn btn-default list_control_icon' name=\"destinations[$x][action]\" alt='".$text['button-delete']."' value='delete'><span class='fas fa-minus'></span></button>";
 			}
 			echo "	</td>\n";
 			echo "</tr>\n";
